refactor: mockData 파일에 있던 타입 별도 파일로 분리
