+++ src/mock/data.ts
@@ -1,24 +1,4 @@
-export interface Store {
-  id: number;
-  name: string;
-  address: string;
-  distance: number;
-  kakaoMapUrl: string;
-  imageUrl: string;
-  rating: number;
-}
-
-export interface Review {
-  id: number;
-  author: {
-    username: string;
-    profileImage: string;
-  };
-  content: string;
-  rating: number;
-  menu: string;
-  updatable: boolean;
-}
+import { Review, Store } from "types/commonTypes";
 
 export const stores: Store[] = [
   {
