+++ src/types/commonTypes.ts
@@ -1,0 +1,21 @@
+export interface Store {
+  id: number;
+  name: string;
+  address: string;
+  distance: number;
+  kakaoMapUrl: string;
+  imageUrl: string;
+  rating: number;
+}
+
+export interface Review {
+  id: number;
+  author: {
+    username: string;
+    profileImage: string;
+  };
+  content: string;
+  rating: number;
+  menu: string;
+  updatable: boolean;
+}
